+++ src/routes/channel/[id]/hoverMenu.svelte
@@ -1,52 +1,13 @@
 <script lang="ts">
 import {IconDots} from "@tabler/icons-svelte";
-import {onMount} from "svelte";
-import { toastStore } from "$lib/store/toast";
-import { repositoryFactory } from "$lib/repositories/RepositoryFactory";
+import { messageDeleteStore } from "$lib/store/messageDelete"
 
-const messageRepository = repositoryFactory.get("message");
+
 
 export let messageId: string = '';
 export let hoverMessageId: string = '';
 
 let isHoverDropDown:boolean = false;
-let deleteModal: HTMLDialogElement;
-
-const deleteMessage = async (messageId :string)=>{
-    console.log("deleting");
-    await messageRepository
-        .deleteMessage(messageId)
-        .then((response) => {
-            console.log("メッセージ削除成功",response);
-            deleteModal.close();
-            toastStore.update((toast) =>{
-                return [
-                    ...toast,
-                    {
-                        message: "メッセージを削除しました。",
-                        type: "success",
-                    }
-                ]
-            })
-        })
-        .catch((response) =>{
-            console.log("メッセージ削除失敗",response);
-            deleteModal.close();
-            toastStore.update((toast) =>{
-                return [
-                    ...toast,
-                    {
-                        message: "メッセージを削除できませんでした。",
-                        type: "error",
-                    },
-                ];
-            });
-        });
-}
-
-onMount(()=>{
-    deleteModal = document.getElementById("deleteModal") as HTMLDialogElement;
-})
 </script>
 
 <div
@@ -73,23 +34,11 @@ onMount(()=>{
             </li>
             <li>
                 <button class="btn btn-sm btn-link hover:text-error"
-                        on:click={() => deleteModal.showModal()}
+                        on:click={() => messageDeleteStore.set(messageId)}
                 >メッセージを削除</button
                 >
             </li>
         </ul>
     </div>
 </div>
-<dialog id="deleteModal" class="modal modal-bottom sm:modal-middle">
-    <div class="modal-box">
-        <h3 class="text-lg font-bold">メッセージを削除</h3>
-        <p class="py-4">メッセージを削除します。よろしいですか?</p>
-        <div class="modal-action">
-            <form method="dialog">
-                <!-- if there is a button in form, it will close the modal -->
-                <button class="btn">キャンセル</button>
-            </form>
-            <button class="btn" on:click={()=> deleteMessage(messageId)}>削除する</button>
-        </div>
-    </div>
-</dialog>
+
